requirements.txt
asyncio
asyncpg
google-api-python-client
google-auth
google-auth-oauthlib
google-cloud
google-cloud-aiplatform
gcsfs
pgvector
PyPDF2